requirements.txt
setuptools>=65.5.1
wheel>=0.38.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
transformers==4.38.2
torch==2.2.1
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
accelerate==0.27.2
bitsandbytes==0.42.0
safetensors==0.4.2
einops==0.7.0
huggingface-hub==0.21.4
flask==3.0.2
flask-cors==4.0.0
openai==1.12.0
numpy==1.26.4
httpx==0.27.0
